html/install.php: 	#  further fixes for url redirection in installer html/modules/installer/xaradmin.php: 	#  further fixes for url redirection in installer html/modules/installer/xartemplates/admin-phase3.xd: 	#  further fixes for url redirection in installer html/includes/xarVar.php: 	#  adding checkbox validation html/includes/xarVar.php: 	#  adding checkbox validation html/includes/xarVar.php: 	#  fixed float validation bounds checking

+++ html/includes/xarVar.php
@@ -347,13 +347,13 @@ class xarVarValidator_float extends xarVarValidator {
         $this->subject = $value;
 
         if (isset($this->parameters[0]) && !empty($this->parameters[0])) {
-            if ($value < (int) $this->parameters[0]) {
+            if ($value < (float) $this->parameters[0]) {
                 return false;
             }
         }
 
         if (isset($this->parameters[1]) && !empty($this->parameters[1])) {
-            if ($value > (int) $this->parameters[1]) {
+            if ($value > (float) $this->parameters[1]) {
                 return false;
             }
         }
